Refactoring to main
